[2911][FIX] account_move_solomon_csv: amount consistency
Before this fix, there were cases where debit and credit balances did not match
when there were multiple analytic accounts from the same analytic plan assigned
to a move line, due to rounding.

This commit fixes the issue by identifying the rounding difference and applying it
to the last row of the move line.